Add support for the VMM Communication exception handler (#VC)
For SEV-ES guests, certain types of VMEXIT events cause a new exception,
VMM Communication exception (#VC), to be taken within the guest. Add
initial support for the #VC handler. The #VC handler requires a GHCB page
in order to handle the exception. During early bootup, a special GHCB page
is used until per-CPU allocations are supported. At that point GHCB usage
is switched over to the per-CPU GHCB structures. The initial support in
the #VC handler will be to report all #VC events as unsupported events to
the hypervisor.

Signed-off-by: Tom Lendacky <thomas.lendacky@amd.com>