Add SEV-ES guest IO read/write support
IO within an SEV-ES guest will generate a non-automatic exit (NAE)
event for an IOIO_PROT event that results in a #VC.

To support IO access within an SEV-ES guest, the instruction must parsed
in order to determine whether it is an IO read or an IO write. The
appropriate register state is set in the GHCB and VMGEXIT processing for
an IOIO_PROT event is performed as documented in the GHCB specification.

Signed-off-by: Tom Lendacky <thomas.lendacky@amd.com>