Add SEV/SEV-ES support for intercepting INVD
The INVD instruction intercept performs emulation. Emulation can't be done
on an SEV or SEV-ES guest because the guest memory is encrypted.

Provide a specific intercept routine for the INVD intercept. Within this
intercept routine, skip the instruction for an SEV guest since it is
emulated as a NOP anyway.