x86/traps: Add SEV-ES guest MSR read/write support
RDMSR and WRMSR within an SEV-ES guest will generate a non-automatic exit
(NAE) event for an MSR_PROT event that results in a #VC.

To support MSR access within an SEV-ES guest, the instruction must parsed
in order to determine whether it is an MSR read or an MSR write. The
appropriate register state is set in the GHCB and VMGEXIT processing for
an MSR_PROT event is performed as documented in the GHCB specification.

Signed-off-by: Tom Lendacky <thomas.lendacky@amd.com>