x86/traps: Add SEV-ES guest DR7 read/write support
DR7 access within an SEV-ES guest will generate a non-automatic exit (NAE)
event for a DR7 read or write event that results in a #VC.

The #VC handler must provide special support to the guest for this as
documented in the GHCB specification. For a DR7 write, the #VC handler
must cache the value and perform standard VMGEXIT processing to notify the
hypervisor of the write. However, the #VC handler must not actually set
the value of the DR7 register. On a DR7 read, the #VC handler must return
the cached value of the DR7 register to the guest. VMGEXIT processing is
not performed for a DR7 register read.

Signed-off-by: Tom Lendacky <thomas.lendacky@amd.com>