x86/mm/encrypt: Fix the effective address calculation in #VC instruction parsing
The code in __default_send_IPI_dest_field was accessing the Local APICs
ICR using instructions with a ModRM.mod field of 0 and a SIB.base of 5.
For example:

0xffffffff81052aae <__default_send_IPI_dest_field+14>:  mov    0xffffffffff5fd300,%eax

The code in vmg_insn_rmdata decoded this to an effective address of 0
which is not correct.  It should be handling this encoding similar to the
way the code in insn-eval.c handles it.

[tl - massaged commit message and tabbing]

Signed-off-by: Doug Covelli <dcovelli@vmware.com>
Signed-off-by: Tom Lendacky <thomas.lendacky@amd.com>