x86/mm: Add initial support for the GHCB
The Guest-Hypervisor Communication Block (GHCB) is used by an SEV-ES guest
to communicate required register state to the hypervisor in order for the
hypervisor to provide required functionality for the guest (e.g. CPUID).

The GHCB follows the format of the VMCB Layout, State Save Area for SEV-ES
in the AMD Architecure Programmer's Manual. The SEV-ES save area is an
extension of the standard VMCB Layout, State Save Area. As such, the
vmcb_save_area struct can be extended for use by the GHCB structure and to
ease implementation of SEV-ES support in KVM in upcoming patches.

Signed-off-by: Tom Lendacky <thomas.lendacky@amd.com>