x86/nmi/64: Add support for NMIs under SEV-ES guest
When running as a guest, the detection of an NMI completing is usually
done by intercepting the IRET instruction. This presents a problem when
running as an SEV-ES guest. Any #VC exception that is taken during NMI
processing will result in an IRET exiting the #VC handler. Since IRETs
are being intercepted, this results in another #VC, which results in
another IRET, etc.

The GHCB specification defines a new behavior for determining the
completion of NMI processing under SEV-ES. The hypervisor will wait for
a VMGEXIT with an NMI-Complete message. At that point, the hypervisor
is free to inject more NMIs. This also allows the guest OS to handle
how it wants to generate the NMI-Complete message in a way that best
works for it.

When exiting an NMI under SEV-ES, the NMI handler will set the Trap Flag
(TF) bit in the RFLAGS register immediately before the IRET instruction.
This will result in the IRET instruction executing and then trapping
into the #VC handler. The #VC handler will see the reason for the trap
being a #DB, issues the NMI-Complete message and exits.

Signed-off-by: Tom Lendacky <thomas.lendacky@amd.com>