KVM: SVM: Clear CR8 intercept for an SEV-ES guest
For SEV-ES guests, the interception of control register write access
is not recommended. Control register interception occurs prior to the
control register being modified and the hypervisor is unable to modify
the control register itself because the register is located in the
encrypted register state.

Clear the intercept for CR8 writes.

Signed-off-by: Tom Lendacky <thomas.lendacky@amd.com>