crypto: ccp: Add support for SEV-ES to PSP driver
In order to provide support for SEV-ES, the hypervisor must provide an
area of memory to the PSP. Once this Trusted Memory Region (TMR) is
provided to the PSP, the contents of this area of memory are no longer
available to the x86.

Update the PSP driver to allocate a 1MB region for the TMR that is 1MB
aligned and then provide it to the PSP on the SEV INIT command.

Signed-off-by: Tom Lendacky <thomas.lendacky@amd.com>