x86/traps: Add SEV-ES guest MMIO read/write support
MMIO within an SEV-ES guest will generate a non-automatic exit (NAE)
event for a nested page fault that results in a #VC.

To support MMIO within an SEV-ES guest, the instruction must parsed in
order to determine whether it is an MMIO read or an MMIO write. The
appropriate register state is set in the GHCB and VMGEXIT processing for
an MMIO event is performed as documented in the GHCB specification.

Co-developed-by: Doug Covelli <dcovelli@vmware.com>
Signed-off-by: Doug Covelli <dcovelli@vmware.com>
Signed-off-by: Tom Lendacky <thomas.lendacky@amd.com>