KVM: SVM: Provide support for launch and run an SEV-ES guest
An SEV-ES guest requires some additional steps to be launched as compared
to an SEV guest:
  - Update MSR_VM_HSAVE_PA to include the encryption bit if SME is active.
  - Alter how some initial register states are set.
  - Choose an SEV-ES ASID value. Only certain ASID values are available
    for SEV-ES guests (configurable through a BIOS/UEFI option).
  - Measure all vCPUs using the LAUNCH_UPDATE_VMSA SEV command after all
    calls to LAUNCH_UPDATE_DATA have been performed but before the call
    to LAUNCH_MEASURE has been performed.

At this point the guest can be run. However, the run sequence is different
for an SEV-ES guest compared to a normal or even an SEV guest. Because the
guest register state is encrypted, it is all saved as part of VMRUN/VMEXIT
and does not require restoring before or saving after a VMRUN instruction.
As a result, all that is required to perform a VMRUN is to save the RBP
and RAX registers, issue the VMRUN and then restore RAX and RBP.

Additionally, certain state is automatically saved and restored with an
SEV-ES VMRUN. As a result certain register states are not required to be
restored upon VMEXIT (e.g. FS, GS, etc.), so only do that is the guest is
not an SEV-ES guest.

Signed-off-by: Tom Lendacky <thomas.lendacky@amd.com>