x86/smpboot: Add support to boot SEV-ES APs
As part of the GHCB specification, the booting of APs under SEV-ES
requires an AP jump table when transitioning from one layer of code to
another (e.g. when going from UEFI to the OS). As a result, each layer
that parks an AP must provide the physical address of an AP jump table
to the next layer using the GHCB MSR.

Upon booting of the kernel, read the GHCB MSR and save the address of
the AP jump table. Under SEV-ES, APs are started using the INIT-SIPI-SIPI
sequence. Before issuing the first SIPI request for an AP, the start eip
is programmed into the AP jump table. Upon issuing the SIPI request, the
AP will awaken and jump to the start eip address programmed into the AP
jump table.

Signed-off-by: Tom Lendacky <thomas.lendacky@amd.com>