x86/paravirt: Per hypervisor GHCB register state support
Add support to allow a hypervisor to provide the GHCB register state that
is required to satisfy a specific hypercall. This allows additional state
beyone the RAX register and the CPL to be included.

As part of overall hypervisor detection support in the kernel, add a
callable function entry to the x86_hyper_runtime struct that will be
called when a hypercall is made and SEV-ES is active. The hypervisor can
register a function that allows it to provide the required register state
in the GHCB that is needed for the hypercall.

Signed-off-by: Tom Lendacky <thomas.lendacky@amd.com>